+++ EdiApi.Client/Types/Parties/PaidFeatureType.cs
@@ -1,0 +1,9 @@
+﻿namespace SkbKontur.EdiApi.Client.Types.Parties
+{
+    public enum PaidFeatureType
+    {
+        Unknown = 0,
+        Module1C = 1,
+        FtpTransport = 2,
+    }
+}