+++ EdiApi.Client/Types/Parties/PaidPeriod.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace SkbKontur.EdiApi.Client.Types.Parties
+{
+    public class PaidPeriod
+    {
+        public DateTime? StartDate { get; set; }
+        public DateTime? EndDate { get; set; }
+    }
+}